backend/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4